Update NuGet package

+++ src/Models/Scratchcard.cs
@@ -2,17 +2,17 @@
 
 public record Scratchcard(
     string Name,
-    int Price,
+    uint Price,
     DateOnly ReleaseDate,
     DateOnly? EndDate,
-    int TotalSupply,
-    Dictionary<int, int> PrizeStructure)
+    uint TotalSupply,
+    Dictionary<uint, uint> PrizeStructure)
 {
     public bool IsAvailable => !EndDate.HasValue || EndDate.Value > DateOnly.FromDateTime(DateTime.Now);
-    public int Jackpot => PrizeStructure.Max(ps => ps.Key);
+    public uint Jackpot => PrizeStructure.Max(ps => ps.Key);
     public decimal ChanceOfWinningToOne => 1 / ChanceOfWinningPercent;
     public decimal ChanceOfWinningPercent => (decimal)PrizeStructure.Sum(ps => ps.Value) / TotalSupply;
-    public decimal ExpectedValue => PrizeStructure.Sum(ps => (long)ps.Key * ps.Value) / TotalSupply;
+    public decimal ExpectedValue => PrizeStructure.Sum(ps => ps.Key * ps.Value) / TotalSupply;
     public decimal ExpectedValuePercent => ExpectedValue / Price;
     public string PathName => Name.Replace(' ', '_');
 }
